requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-forms==2.0
django-phone-field==1.8.1
mysqlclient==2.2.0
sqlparse==0.4.4
tzdata==2023.3
